+++ app/javascript/components/widget/components/widget-map-list/styles.scss
@@ -1,6 +1,6 @@
 @import '~styles/settings.scss';
 
-.c-widget-trase-commodities {
+.c-widget-map-list {
   display: flex;
   flex-direction: column;
   justify-content: space-between;
