--- app/javascript/components/widgets/land-use/economic-impact/actions.js
@@ -1,34 +1,0 @@
-import uniq from 'lodash/uniq';
-import sortBy from 'lodash/sortBy';
-
-import { getFAOEcoLive } from 'services/forest-data';
-
-export default ({ params }) =>
-  getFAOEcoLive(params.token).then(response => {
-    const { rows } = response.data;
-    const { adm0, year } = params;
-    const years = sortBy(
-      uniq(
-        rows
-          .filter(
-            d =>
-              d.country === adm0 &&
-              d.year !== 9999 &&
-              d.usdrev !== null &&
-              d.usdexp !== null &&
-              d.usdexp !== ''
-          )
-          .map(d => ({ label: d.year, value: d.year }))
-      )
-    );
-    const payload = {
-      data: rows || {},
-      options: {
-        years
-      },
-      settings: {
-        year: years.includes(year) ? year.value : years[0].value
-      }
-    };
-    return payload;
-  });
