--- app/javascript/components/widgets/land-use/economic-impact/config.js
@@ -1,139 +1,0 @@
-export default {
-  widget: 'economicImpact',
-  title: 'Economic Impact of forests in {location}',
-  large: true,
-  categories: ['land-use'],
-  types: ['country'],
-  admins: ['adm0'],
-  options: {
-    years: true,
-    units: ['net_usd', 'net_perc']
-  },
-  dataType: 'fao',
-  colors: 'employment',
-  metaKey: 'widget_economic_impact',
-  sortOrder: {
-    landUse: 2
-  },
-  sentence:
-    'According to the FAO, the forestry sector contributed a net {value} to the economy in {year}, which is approximately {percentage} of {location} GDP.',
-  whitelists: {
-    adm0: [
-      'AGO',
-      'ALB',
-      'ARG',
-      'ARM',
-      'BEL',
-      'BEN',
-      'BFA',
-      'BGD',
-      'BGR',
-      'BLM',
-      'BLR',
-      'BRA',
-      'BRN',
-      'BTN',
-      'BWA',
-      'CHL',
-      'CHN',
-      'CRI',
-      'CUB',
-      'CYP',
-      'DNK',
-      'DOM',
-      'DZA',
-      'EGY',
-      'FIN',
-      'FJI',
-      'FLK',
-      'FRA',
-      'FRO',
-      'GAB',
-      'GBR',
-      'GEO',
-      'GIB',
-      'GLP',
-      'GMB',
-      'GNQ',
-      'GRL',
-      'GTM',
-      'GUF',
-      'GUY',
-      'HRV',
-      'HUN',
-      'IDN',
-      'IND',
-      'IRN',
-      'ISL',
-      'JAM',
-      'JPN',
-      'KEN',
-      'KGZ',
-      'KHM',
-      'KIR',
-      'KOR',
-      'LAO',
-      'LBN',
-      'LCA',
-      'LKA',
-      'LTU',
-      'LVA',
-      'MAR',
-      'MCO',
-      'MEX',
-      'MLI',
-      'MMR',
-      'MNE',
-      'MNG',
-      'MRT',
-      'MTQ',
-      'MUS',
-      'MWI',
-      'MYS',
-      'MYT',
-      'NER',
-      'NPL',
-      'NRU',
-      'NZL',
-      'PAN',
-      'PER',
-      'PHL',
-      'PNG',
-      'POL',
-      'PRY',
-      'PYF',
-      'QAT',
-      'REU',
-      'RUS',
-      'RWA',
-      'SDN',
-      'SEN',
-      'SGP',
-      'SJM',
-      'SLE',
-      'SMR',
-      'SRB',
-      'SUR',
-      'SVK',
-      'SVN',
-      'SWE',
-      'SYC',
-      'SYR',
-      'TGO',
-      'TKL',
-      'TON',
-      'TTO',
-      'TUN',
-      'TUR',
-      'TZA',
-      'URY',
-      'USA',
-      'UZB',
-      'VAT',
-      'VEN',
-      'VUT',
-      'WLF',
-      'ZWE'
-    ]
-  }
-};
