--- app/javascript/components/widgets/land-use/economic-impact/settings.js
@@ -1,3 +1,0 @@
-export default {
-  unit: 'net_usd'
-};
