--- app/javascript/components/widgets/land-use/forestry-employment/actions.js
@@ -1,4 +1,0 @@
-import { getFAOEcoLive } from 'services/forest-data';
-
-export default ({ params }) =>
-  getFAOEcoLive(params.token).then(response => response.data.rows);
