--- app/javascript/components/widgets/land-use/forestry-employment/settings.js
@@ -1,4 +1,0 @@
-export default {
-  year: 2010,
-  unit: ''
-};
