--- app/javascript/components/widgets/land-use/trase-commodities/settings.js
@@ -1,6 +1,0 @@
-export default {
-  unit: 't',
-  pageSize: 5,
-  page: 0,
-  commodity: 'SOY'
-};
